Minor fixes

+++ src/sprite_image.c
@@ -149,7 +149,7 @@ sd_rgba_image* sd_sprite_image_decode(sd_sprite_image *img, sd_palette *pal, int
                         rgba->data[pos+1] = (uint8_t)pal->data[b][1];
                         rgba->data[pos+2] = (uint8_t)pal->data[b][2];
                     }
-                    rgba->data[pos+3] = 255; // fully opaque
+                    rgba->data[pos+3] = (uint8_t)255; // fully opaque
                     i++; // we read 1 byte
                     x++;
                     data--;
